pyproject.toml
[tool.poetry]
name = "TelQTelecomAPI"
version = "0.1.0"
description = "Python SDK for TelQ Telecom API"
authors = ["Tuvshinbayar Davaa <tuvshinbayar.davaa@telqtele.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.20.0"
requests = "^2.27.1"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
